atualizando estrutura do projeto - login e oportunidades
