+++ frontend/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
   <head>
     <meta charset="UTF-8" />
-    <link rel="icon" type="image/svg+xml" href="/vite.svg" />
+    <link rel="icon" type="image/x-icon" sizes="192x192" href="public/public/WeUP.ico" />
     <meta name="viewport" content="width=device-width, initial-scale=1.0" />
     <title>WeUnite</title>
   </head>
