+++ frontend/src/App.jsx
@@ -2,24 +2,24 @@ import { Container } from "@chakra-ui/react";
 import PostPage from "./pages/PostPage";
 import UserPage from "./pages/UserPage";
 import { Navigate, Route, Routes } from "react-router-dom";
-import AuthPage from "./pages/AuthPage";
+import AuthPage from "./pages/login/AuthPage";
 import HomePage from "./pages/HomePage";
 import UpdateProfilePage from "./pages/UpdateProfilePage";
 import CreatePost from "./components/CreatePost";
-import CreateOportunity from "./components/CreateOportunity";
+import CreateOportunity from "./components/oportunities/CreateOportunity";
 import ChatPage from "./pages/ChatPage";
-import OportunityPage from "./pages/OportunityPage";
-import ForgotPasswordPage from "./pages/ForgotPasswordPage";
-import ResetPasswordPage from "./pages/ResetPasswordPage";
-import VerifyResetCodePage from "./pages/VerifyResetCodePage";
-import VerifyEmailPage from "./pages/VerifyEmailPage";
-import OportunityDetailsPage from "./pages/OportunityDetailsPage";
-import SavedOportunitiesPage from "./pages/SavedOportunitiesPage";
+import OportunityPage from "./pages/oportunity/OportunityPage";
+import ForgotPasswordPage from "./pages/login/ForgotPasswordPage";
+import ResetPasswordPage from "./pages/login/ResetPasswordPage";
+import VerifyResetCodePage from "./pages/login/VerifyResetCodePage";
+import VerifyEmailPage from "./pages/login/VerifyEmailPage";
+import OportunityDetailsPage from "./pages/oportunity/OportunityDetailsPage";
+import SavedOportunitiesPage from "./pages/oportunity/SavedOportunitiesPage";
 import { useRecoilValue } from "recoil";
 import userAtom from "./atoms/userAtom";
 import MyApplicationsPage from "./pages/MyApplicationsPage";
 import ClubOpportunitiesPage from "./pages/ClubOpportunitiesPage";
-import OpportunityApplicationsPage from "./pages/OpportunityApplicationsPage";
+import OpportunityApplicationsPage from "./pages/oportunity/OpportunityApplicationsPage";
 
 function App() {
     const user = useRecoilValue(userAtom);
