+++ frontend/src/components/login/CompanySignUpCard.jsx
@@ -18,11 +18,11 @@ import {
     RequiredIndicator,
 } from '@chakra-ui/react';
 import { useSetRecoilState } from 'recoil';
-import useShowToast from '../hooks/useShowToast';
+import useShowToast from '../../hooks/useShowToast';
 
 // Importe seus átomos e hooks aqui
-import authScreenAtom from '../atoms/authAtom';
-import TermsCheckbox from './TermsCheckBox';
+import authScreenAtom from '../../atoms/authAtom';
+import TermsCheckbox from '../TermsCheckBox';
 
 const validateCNPJ = (cnpj) => {
     cnpj = cnpj.replace(/\D/g, ''); // Remove caracteres não numéricos
