+++ frontend/src/components/login/PasswordStrengthMeter.jsx
