+++ frontend/src/pages/login/VerifyResetCodePage.jsx
@@ -1,21 +1,18 @@
-import { Button, Flex, FormControl, Heading, Input, Spinner, Text, useColorMode } from "@chakra-ui/react"
+import { Button, Flex, FormControl, Heading, Input, Text, useColorMode } from "@chakra-ui/react"
 import { useEffect, useRef, useState } from "react";
+import useShowToast from "../../hooks/useShowToast";
+import { useAuthStore } from "../../store/authStore";
 import { useNavigate } from "react-router-dom";
-import useShowToast from "../hooks/useShowToast";
-import { useAuthStore } from "../store/authStore";
-import { useSetRecoilState } from "recoil";
-import userAtom from "../atoms/userAtom";
 
-
-const VerifyEmailPage = () => {
+const VerifyResetCodePage = () => {
     const [code, setCode] = useState(["", "", "", "", "", ""]);
     const inputRefs = useRef([]);
-    const navigate = useNavigate();
     const showToast = useShowToast();
-    const setUser = useSetRecoilState(userAtom);
+    const navigate = useNavigate();
 
     const { colorMode } = useColorMode();
-    const { isLoading, verifyEmail } = useAuthStore();
+
+    const { isLoading, verifyResetCode } = useAuthStore();
 
     const handleChange = (index, value) => {
         const newCode = [...code];
@@ -53,19 +50,20 @@ const VerifyEmailPage = () => {
         e.preventDefault();
         const verificationCode = code.join("");
         try {
-            const res = await verifyEmail(verificationCode);
-            if (res.success) {
-                // Agora podemos salvar o usuário no localStorage, pois ele está verificado
-                localStorage.setItem("user-threads", JSON.stringify(res.user));
-                setUser(res.user);
-                showToast("Sucesso", "Conta ativada com sucesso!", "success");
-                navigate("/");
-            } else {
-                throw new Error(res.message || "Falha na verificação");
-            }
+            await verifyResetCode(verificationCode);
+            showToast("Successo", "Código verificado com sucesso!", "success");
+            navigate(`/reset-password/${verificationCode}`);
         } catch (error) {
-            showToast("Erro", "Código inválido ou expirado", "error");
-            console.log(error);
+            // Reset the code inputs
+            setCode(["", "", "", "", "", ""]);
+            // Focus on the first input
+            inputRefs.current[0].focus();
+            // Show error toast
+            showToast(
+                "Erro", 
+                error?.response?.data?.message || "Código inválido ou expirado", 
+                "error"
+            );
         }
     };
 
@@ -76,26 +74,39 @@ const VerifyEmailPage = () => {
         }
     }, [code]);
 
-
     return (
-        <Flex width={"100vw"} height={"100vh"} alignItems={"center"} justifyContent={"center"}
+        <Flex 
+            width={"100vw"} 
+            height={"100vh"} 
+            alignItems={"center"} 
+            justifyContent={"center"}
             backgroundColor={colorMode === "dark" ? "#0A0A0A" : "gray.200"}
         >
-            <Flex flexDir={"column"} alignItems={"center"} justifyContent={"center"}
-                backgroundColor={colorMode === "dark" ? "black" : "gray.100"} //Mudei
-                borderRadius="lg" shadow="md" //Mudei
-                p={10}
-                border={colorMode ==="dark" ? '1px solid #101010' : ''}>
-
-                <Flex flexDir={"column"} gap={3} alignItems={"center"} justifyContent={"center"} mb={1} p={5}>
-                    <Heading>Verifique seu email</Heading>
-                    <Text color={"#03C03C"}>Insira o código enviado para você</Text>
+            <Flex 
+                flexDir={"column"} 
+                alignItems={"center"} 
+                justifyContent={"center"}   
+                backgroundColor={colorMode === "dark" ? "black" : "gray.100"}
+                borderRadius="lg" 
+                shadow="md"
+                p={10} 
+                border={colorMode === "dark" ? "1px solid #101010" : ""}
+            >
+                <Flex 
+                    flexDir={"column"} 
+                    gap={3} 
+                    alignItems={"center"} 
+                    justifyContent={"center"} 
+                    mb={3} 
+                    p={5}
+                >
+                    <Heading>Verifique seu código</Heading>
+                    <Text color={'#03C03C'}>Insira o código enviado para você</Text>
                 </Flex>
 
                 <Flex gap={2}>
                     <FormControl gap={2}>
                         {code.map((digit, index) => (
-
                             <Input
                                 key={index}
                                 ref={(el) => (inputRefs.current[index] = el)}
@@ -108,20 +119,26 @@ const VerifyEmailPage = () => {
                                 height={12}
                                 textAlign={"center"}
                                 fontWeight={"bold"}
-                                border={`1px solid ${digit ? "#03C03C" : "#343434"}`} // Change border color based on input value
                                 borderRadius={5}
                                 mx={2}
-                                focusBorderColor={digit ? "#03C03C" : "#343434"} // Optional: Change border color on focus
+                                focusBorderColor={digit ? "#03C03C" : "#343434"}
+                                border={`1px solid ${digit ? "#03C03C" : "#343434"}`}
                             />
-
                         ))}
                     </FormControl>
                 </Flex>
 
-                <Button type="submit" variant={"outline"} disabled={isLoading || code.some((digit) => !digit)} mt={5}>	{isLoading ? <Spinner size="sm" color="#03C03C" /> : "Cadastrar"}</Button>
+                <Button 
+                    type="submit" 
+                    variant={"outline"} 
+                    disabled={isLoading || code.some((digit) => !digit)} 
+                    mt={5}
+                >
+                    {isLoading ? "Verificando..." : "Verificar Email"}
+                </Button>
             </Flex>
         </Flex>
-    )
-}
+    );
+};
 
-export default VerifyEmailPage
+export default VerifyResetCodePage;
