+++ frontend/src/pages/oportunity/OportunityDetailsPage.jsx
@@ -2,11 +2,11 @@ import { useState, useEffect } from "react";
 import { Box, Button, Flex, HStack, Icon, Image, Tag, Text, useColorModeValue, VStack,useColorMode } from "@chakra-ui/react"
 import { useParams } from "react-router-dom";
 import { useRecoilState, useRecoilValue } from "recoil";
-import useShowToast from "../hooks/useShowToast";
-import useGetUserProfile from "../hooks/useGetUserProfile";
-import postsAtom from "../atoms/postsAtom";
-import userAtom from "../atoms/userAtom";
-import Header from "../components/Header";
+import useShowToast from "../../hooks/useShowToast";
+import useGetUserProfile from "../../hooks/useGetUserProfile";
+import postsAtom from "../../atoms/postsAtom";
+import userAtom from "../../atoms/userAtom";
+import Header from "../../components/Header";
 import { FaMapMarkerAlt, FaClock } from 'react-icons/fa';
 
 const OportunityDetailsPage = () => {
